requirements.txt
psycopg2-binary
pytest
selenium
allure-pytest
requests